feat: Create a file for each function.
