+++ src/data/add_struct_redirect.c
@@ -1,31 +1,43 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   free_redirects.c                                   :+:      :+:    :+:   */
+/*   add_struct_redirect.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/16 19:27:06 by tkondo            #+#    #+#             */
-/*   Updated: 2025/02/22 22:39:10 by miyuu            ###   ########.fr       */
+/*   Created: 2025/02/22 22:35:12 by miyuu             #+#    #+#             */
+/*   Updated: 2025/02/22 22:35:17 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
 /*
- * Function:
+ * Function:add_struct_redirect
  * ----------------------------
- * free memory of redirects
+ * Fill t_redirect with data.
+ *
+ * char *path = File path.
+ * int type = The type of redirect (input, output, append).
  */
-void	free_redirects(t_redirect *reds)
+void	add_struct_redirect(t_redirect **reds, int type, char *path)
 {
+	t_redirect	*new;
 	t_redirect	*tmp;
 
-	while (reds != NULL)
+	new = malloc(sizeof(t_redirect));
+	if (!new)
+		return ;
+	new->path = ft_strdup(path);
+	new->red_type = type;
+	new->next = NULL;
+	if (*reds == NULL)
+		*reds = new;
+	else
 	{
-		free((char *)reds->path);
-		tmp = reds;
-		reds = reds->next;
-		free(tmp);
+		tmp = *reds;
+		while (tmp->next)
+			tmp = tmp->next;
+		tmp->next = new;
 	}
 }
