+++ src/data/fill_words.c
@@ -1,0 +1,50 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   fill_words.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/22 22:36:06 by miyuu             #+#    #+#             */
+/*   Updated: 2025/02/22 22:36:27 by miyuu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <minishell.h>
+
+/*
+ * Function:fill_words
+ * ----------------------------
+ * Returns an array only strings without redirects.
+ * ToDO:norminetteエラー
+ */
+char	**fill_words(char **src, int wc)
+{
+	char	**dst;
+	int		i;
+	int		j;
+
+	i = 0;
+	j = 0;
+	dst = (char **)malloc(sizeof(char *) * (wc + 1));
+	if (!dst)
+		return (NULL);
+	while (src[i])
+	{
+		if (has_redirect(src[i]) == true && src[i + 1])
+			i++;
+		else
+		{
+			dst[j] = ft_strdup(src[i]);
+			if (!dst[j])
+			{
+				free_words(dst);
+				return (NULL);
+			}
+			j++;
+		}
+		i++;
+	}
+	dst[j] = NULL;
+	return (dst);
+}
